Die Profil-Formulare für E-Mail-Adresse und Passwort nehmen bei abgeschaltetem Passwort-Login die per Passkey bestätigte Sitzung als Nachweis — bisher verlangten sie ein Passwort, das dort nichts mehr gilt, und Enter lief am Dialog vorbei; die Frist dafür ist dieselbe wie die der Middleware, auch wenn sie aus der Umgebung als Text kommt.

+++ config/auth.php
@@ -114,6 +114,6 @@ return [
     |
     */
 
-    'password_timeout' => env('AUTH_PASSWORD_TIMEOUT', 10_800),
+    'password_timeout' => (int) env('AUTH_PASSWORD_TIMEOUT', 10_800),
 
 ];
